feat(linux_can)!: Implement CAN frame filtering
Try to use the kernel for filtering, if not, filter in userspace.
If multiple different filters are requested per socket, try to combine
them into one that minimally covers both individual filters, so we have
to check as few frames as possible in userspace.

Add CanFilterNotRepresentableException, for when a dart CanFilter
instance is not representable as a native SocketCAN filter list. Allows
us to fall back to userspace filtering in that case.

Rework the CAN frame class hierarchy. Add 4 abstract (sealed) marker
classes:
- CanBaseFrame, CanExtendedFrame
- CanDataFrame, CanRemoteFrame

The 4 concrete implementations of CanFrame (CanStandardDataFrame, etc)
are now direct descendants of `CanFrame` and implement the marker
classes correspondingly. Remove CanFrame.error, errors should not be
disguised as CAN frames.

Allows us to exhaustively switch over whatever property we choose, i.e:
```
final x = switch (frame) {
  CanDataFrame(:final data) => ...,
  CanRemoteFrame _ => ...,
}
```

or:
```
final x = switch (frame) {
  CanBaseFrame _ => ...,
  CanExtendedFrame _ => ...,
}
```

Make [CanError] a class (implementing Exception), add all specific
CAN error variants, and add human-readable descriptions for these
errors.

Add CanFilter.{whitelist,blacklist} shortcuts.

Combine CanFilter.rules and CanFilter.ruleCombinator into a single
getter that returns a record. This means we only have a single place
to throw a CanFilterNotRepresentableException, results in more
consistent behaviour.

Add CanFilter.matches for manually checking if a CanFilter matches a frame.

Try to not use the kernel internal buffering. Set the error and frame
filter to none on CAN socket open so no frames are buffered in the
kernel. Only set it to actual, satisfiable filters when we're waiting
for a frame.

Rename `CanSocket.read() => CanSocket.receiveSingle()` and change the
semantics to be in line with what was written above.

Rename `CanSocket.frames => CanSocket.receive()`.

`read` implies the action is fast & non-blocking, which was true before,
but now that we avoid the in-kernel buffer, read (/ receiveSingle) is
actually blocking and starts waiting for a frame to arrive. `receive` is
the more appropriate name IMO. However, we also renamed `frames` to
`receive()`, so use `receiveSingle()` for receiving a single frame.

`frames` needed to become a full method so we can have a filter and
error reporting parameter, ==> `receive()`.

linux_can_test_app: Make sure all loopback tests do the `receive()`
before the `send()`.